Ajout fonction set + Fix bug
J'ai rajouter une fonction qui permet d'écrire une fonction dans un fichier csv. De plus j'ai résolu un bug de type dans equ.c, le type de division était int mais il fallait un double

+++ include/equ.h
@@ -13,6 +13,6 @@ double **GetMatrixFromAugmentedMatrix(double **matrix, int size);
 double **GetResultFromAugmentedMatrix(double **matrix, int size);
 
 // Free
-int FreeMatrix(int **augmented_matrix, int size);
+int FreeMatrix(double **augmented_matrix, int size);
 
 #endif // __EQU_H__
